# swagger.yaml
openapi: 3.0.3
info:
  title: E-Commerce API — Testing Practice
  version: 1.0.0
  description: |
    Full-featured e-commerce REST API built for API testing practice.
    Covers auth flows, CRUD, pagination (offset + cursor), caching (ETag/304),
    file uploads, BOLA/IDOR, rate limiting, fault simulation, idempotency,
    RFC 7807 errors, state machines, and role-based access.

    ## Seeded credentials (reset anytime via `POST /v1/seed/reset`)
    | Role  | Email             | Password   |
    |-------|-------------------|------------|
    | admin | admin@test.com    | Admin123!  |
    | user  | user@test.com     | User123!   |
    | user  | user2@test.com    | User123!   |

    ## Notable testing hooks
    - `?_delay=3000` — delays response by N ms (max 10 000)
    - `?_fail=503`   — returns a simulated error with that status code
    - `Idempotency-Key` header on `POST /orders` — prevents duplicate orders on retry
    - `If-None-Match` header on GET product(s) — returns 304 when data unchanged
servers:
  - url: http://localhost:3000/v1
    description: Local development
  - url: https://YOUR-APP.onrender.com/v1
    description: Production (update after deploying to Render)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Problem:
      type: object
      description: RFC 7807 Problem Details
      required: [type, title, status, detail, instance]
      properties:
        type:     { type: string, example: "http://localhost:3000/errors/not-found" }
        title:    { type: string, example: "Not Found" }
        status:   { type: integer, example: 404 }
        detail:   { type: string, example: "Product with id 99 does not exist" }
        instance: { type: string, example: "/v1/products/99" }
...
    Product:
      type: object
      properties:
        id:          { type: integer, example: 1 }
        name:        { type: string,  example: "Smartphone Pro X" }
        description: { type: string,  example: "Latest flagship smartphone" }
        price:       { type: number,  format: float, example: 999.99 }
        stock:       { type: integer, example: 50 }
        categoryId:  { type: integer, example: 1 }
        imageUrl:    { type: string, nullable: true, example: "/v1/products/1/image" }
        createdAt:   { type: string,  format: date-time }
...
    OffsetPagination:
      type: object
      properties:
        type:       { type: string, example: offset }
        page:       { type: integer }
        limit:      { type: integer }
        total:      { type: integer }
        totalPages: { type: integer }
        hasNext:    { type: boolean }
        hasPrev:    { type: boolean }
    CursorPagination:
      type: object
      properties:
        type:       { type: string, example: cursor }
        limit:      { type: integer }
        nextCursor: { type: integer, nullable: true }
        hasNext:    { type: boolean }
  responses:
    Unauthorized:
      description: Unauthorized — missing, expired, or invalid token
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Forbidden — insufficient role
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    UnprocessableEntity:
      description: Validation error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
paths:
...
  /products:
    get:
      summary: List products
      description: Supports filtering, sorting, full-text search, offset pagination, and cursor pagination. Supports ETag caching and fault simulation.
      tags: [Products]
      parameters:
        - { name: categoryId,    in: query,  schema: { type: integer }, description: Filter by category }
        - { name: search,        in: query,  schema: { type: string  }, description: Full-text search on name + description }
        - { name: sortBy,        in: query,  schema: { type: string, enum: [id, name, price, stock, createdAt], default: id } }
        - { name: order,         in: query,  schema: { type: string, enum: [asc, desc], default: asc } }
        - { name: page,          in: query,  schema: { type: integer, default: 1 }, description: Offset pagination page }
        - { name: limit,         in: query,  schema: { type: integer, default: 10 } }
        - { name: cursor,        in: query,  schema: { type: integer }, description: Cursor pagination — pass last seen product id. Overrides page. }
        - { name: _delay,        in: query,  schema: { type: integer }, description: Simulate slow response (ms, max 10000) }
        - { name: _fail,         in: query,  schema: { type: integer }, description: Simulate error response (e.g. 503) }
        - { name: If-None-Match, in: header, schema: { type: string  }, description: ETag from previous response — returns 304 if data unchanged }
      responses:
        '200':
          description: Product list
          headers:
            ETag:          { schema: { type: string } }
            Cache-Control: { schema: { type: string } }
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:       { type: array, items: { $ref: '#/components/schemas/Product' } }
                  pagination:
                    oneOf:
                      - { $ref: '#/components/schemas/OffsetPagination' }
                      - { $ref: '#/components/schemas/CursorPagination' }
        '304': { description: Not Modified — use cached data }
    post:
      summary: Create product (admin)
      tags: [Products]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, price, stock, categoryId]
              properties:
                name:        { type: string }
                description: { type: string }
                price:       { type: number, example: 49.99 }
                stock:       { type: integer, example: 100 }
                categoryId:  { type: integer, example: 1 }
      responses:
        '201':
          headers: { Location: { schema: { type: string } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/Product' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /products/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      summary: Get product by ID
      tags: [Products]
      parameters:
        - { name: _delay, in: query, schema: { type: integer } }
        - { name: _fail,  in: query, schema: { type: integer } }
        - { name: If-None-Match, in: header, schema: { type: string } }
      responses:
        '200':
          headers:
            ETag:          { schema: { type: string } }
            Cache-Control: { schema: { type: string } }
          content: { application/json: { schema: { $ref: '#/components/schemas/Product' } } }
        '304': { description: Not Modified }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Full replace of product (admin) — omitted fields revert to null/default
      tags: [Products]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, price, stock, categoryId]
              properties:
                name:        { type: string }
                description: { type: string }
                price:       { type: number }
                stock:       { type: integer }
                categoryId:  { type: integer }
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/Product' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Partial update of product (admin) — only supplied fields change
      tags: [Products]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                name:        { type: string }
                description: { type: string }
                price:       { type: number }
                stock:       { type: integer }
                categoryId:  { type: integer }
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/Product' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Delete product (admin)
      tags: [Products]
      security:
        - bearerAuth: []
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /products/{id}/image:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    post:
      summary: Upload product image (admin) — multipart/form-data, max 5MB, JPEG/PNG/WebP
      tags: [Products]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image: { type: string, format: binary }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  imageUrl: { type: string }
                  filename: { type: string }
                  size:     { type: integer }
                  mimeType: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { description: Image exceeds 5 MB }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    get:
      summary: Serve product image
      tags: [Products]
      responses:
        '200':
          description: Image binary
          content:
            image/jpeg: {}
            image/png:  {}
            image/webp: {}
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Cart ────────────────────────────────────────────────────────────────────
...